error handling

+++ backend/flaskr/__init__.py
@@ -10,19 +10,11 @@ from models import setup_db, Question, Category, db
 QUESTIONS_PER_PAGE = 10
 
 def create_app(test_config=None):
-  # create and configure the app
     app = Flask(__name__)
     setup_db(app)
-      # @TODO: Set up CORS. Allow '*' for origins. Delete the sample route after completing the TODOs
     CORS(app, resources={r'/*/api/*': {'origins': '*'}})
 
 
-
-
-
-  # '''
-  # @TODO: Use the after_request decorator to set Access-Control-Allow
-  # '''
     @app.after_request
     def after_request(response):
         response.headers.add('Access_Control_Allow_Headers', 'Content-Type, Authorization, True')
@@ -30,11 +22,6 @@ def create_app(test_config=None):
 
         return response
 
-      # '''
-      # @TODO:
-      # Create an endpoint to handle GET requests
-      # for all available categories.
-      # '''
     @app.route('/categories', methods=['GET'])
     def categories():
         cats=Category.query.all()
@@ -47,11 +34,7 @@ def create_app(test_config=None):
           'categories': all_categories
           })
 
-      # @TODO:
-      # Create an endpoint to handle GET requests for questions,
-      # including pagination (every 10 questions).
-      # This endpoint should return a list of questions,
-      # number of total questions, current category, categories.
+
 
     @app.route('/questions')
     def questions():
@@ -73,16 +56,15 @@ def create_app(test_config=None):
         'categories': all_categories,
         # 'currentCategory':
         })
-  # TEST: At this point, when you start the application
-  # you should see questions and categories generated,
-  # ten questions per page and pagination at the bottom of the screen for three pages.
-  # Clicking on the page numbers should update the questions.
-  # @TODO:
-  # Create an endpoint to DELETE question using a question ID.
+
+
 
     @app.route('/questions/<id>', methods=['DELETE'])
     def delete(id):
        question = Question.query.filter_by(id = id).first()
+       print(question)
+       if not question:
+           abort(422)
        try:
           db.session.delete(question)
           db.session.commit()
@@ -95,10 +77,7 @@ def create_app(test_config=None):
           'success': True
           })
 
-  # @TODO:
-  # Create an endpoint to POST a new question,
-  # which will require the question and answer text,
-  # category, and difficulty score.
+
     @app.route('/questions', methods = ['POST'])
     def add_question():
         body=request.get_json()
@@ -119,9 +98,7 @@ def create_app(test_config=None):
             db.session.close()
             return jsonify({ "success":True})
 
-  # TEST: When you submit a question on the "Add" tab,
-  # the form will clear and the question will appear at the end of the last page
-  # of the questions list in the "List" tab.
+
     @app.route('/questionSearch', methods=['POST'])
     def search():
         body = request.get_json()
@@ -135,27 +112,20 @@ def create_app(test_config=None):
           'totalQuestions': len(total_questions)
           #'currentCategory': categories
         })
-  #
-  # '''
-  # @TODO:
-  # Create a POST endpoint to get questions based on a search term.
-  # It should return any questions for whom the search term
-  # is a substring of the question.
+
     @app.route('/categories/<int:id>/questions', methods = ['GET'])
     def filter_questions(id):
         questions = Question.query.filter_by(category = id).all()
         formated_questions = [question.format() for question in questions]
+        print(questions)
+        if not questions:
+            abort(404)
         return jsonify({
         'success': True,
         'totalQuestions': len(questions),
         'questions': formated_questions
         })
-  # @TODO:
-  # Create a GET endpoint to get questions based on category.
-  #
-  # TEST: In the "List" tab / main screen, clicking on one of the
-  # categories in the left column will cause only questions of that
-  # category to be shown.
+
 
     @app.route('/quizzes', methods = ['POST'])
     def quizzes():
@@ -171,22 +141,25 @@ def create_app(test_config=None):
             "success": True,
             "question": quiz_question
         })
-  # @TODO:
-  # Create a POST endpoint to get questions to play the quiz.
-  # This endpoint should take category and previous question parameters
-  # and return a random questions within the given category,
-  # if provided, and that is not one of the previous questions.
-  #
-  # TEST: In the "Play" tab, after a user selects "All" or a category,
-  # one question at a time is displayed, the user is allowed to answer
-  # and shown whether they were correct or not.
-  # '''
-  #
-  # '''
+
   # @TODO:
   # Create error handlers for all expected errors
   # including 404 and 422.
   # '''
+    @app.errorhandler(404)
+    def not_found(error):
+        return jsonify({
+        "success": False,
+        "error": 404,
+        "message":"resource not found"
+        })
+    @app.errorhandler(422)
+    def not_found(error):
+        return jsonify({
+        "success": False,
+        "error": 422,
+        "message":"resource not found"
+        })
